Refactor: 도메인 단위로 패키지 관리
