+++ src/main/java/team18/team18_be/contract/dto/request/ContractRequest.java
@@ -1,4 +1,4 @@
-package team18.team18_be.dto;
+package team18.team18_be.contract.dto.request;
 
 import jakarta.validation.constraints.Pattern;
